separando as funções em uma pasta functions
