+++ pokedex/src/App.js
@@ -8,7 +8,6 @@ import axios from 'axios';
 import { useState, useEffect, createContext } from 'react';
 
 export const GlobalContext = createContext()
-// GLOBAL STATE SÓ FOI COLOCADO NO HOME
 
 function App() {
   const [pokemonlist, setPokemonlist] = useState([]);
@@ -31,52 +30,14 @@ function App() {
       });
   };
 
-  const buildImgUrl = (url) => {
-    const id = url.split("/");
-    const idx = id.length - 2;
-    const imgUrl = `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${id[idx]}.png`;
-
-    return imgUrl;
-  };
-
-  const getPokemonId = (url) => {
-    const id = url.split("/");
-    const idx = id.length - 2;
-    const pokemonId = id[idx];
-    return pokemonId;
-  };
-
-  const addToPokedex = (pokemon) => {
-    setPokedex([...pokedex, pokemon]);
-    const newPokemonList = pokemonlist.filter((p) => p.name !== pokemon.name);
-    setPokemonlist(newPokemonList);
-  };
-
-  const removeFromPokedex = (pokemon) => {
-    const newPokedex = pokedex.filter((p) => p.name !== pokemon.name);
-    setPokedex(newPokedex);
-    const newPokemonList = [...pokemonlist, pokemon];
-    setPokemonlist(newPokemonList);
-  };
-
   return (
 
-    <GlobalContext.Provider value = {{pokemonlist, pokedex}}>
+    <GlobalContext.Provider value = {{pokemonlist, setPokemonlist, pokedex, setPokedex}}>
       <BrowserRouter>
         <Header />
         <Routes>
-          <Route path = "/" element = {<Home
-          // pokemonlist={pokemonlist}
-          buildImgUrl={buildImgUrl}
-          getPokemonId={getPokemonId}
-          addToPokedex={addToPokedex}
-          />} />
-          <Route path = "/pokedex" element = {<Pokedex 
-          buildImgUrl={buildImgUrl} 
-          pokedex={pokedex}
-          removeFromPokedex={removeFromPokedex}
-          pokemonlist={pokemonlist}
-          getPokemonId={getPokemonId} />} />          
+          <Route path = "/" element = {<Home/>} />
+          <Route path = "/pokedex" element = {<Pokedex/>} />          
           <Route path = "/details/:id" element = {<PokemonDetails />} />          
         </Routes>
       </BrowserRouter>
